separate files and restructure tests
