+++ scheduling/storage_providers/memory_test.go
@@ -1,0 +1,19 @@
+package scheduling
+
+import (
+	"testing"
+)
+
+func TestStorageProviderMemory(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping integration tests")
+	}
+
+	storage := NewMemoryStorageProvider()
+
+	test := StorageProviderTest{
+		t:       t,
+		storage: storage,
+	}
+	test.Run()
+}
